refactor: move gesture detector to util directory
